+++ packages/tswebextension/src/lib/common/storage/extension-storage.ts
@@ -1,30 +1,29 @@
-import { type Storage } from 'webextension-polyfill';
-
 import { PersistentValueContainer } from './persistent-value-container';
+import type { ExtendedStorageInterface } from './core/storage-interface';
 
 /**
- * API for storing persistent key-value data with debounced sync with the specified webextension storage key.
- * Webextension storage synchronization described in the {@link PersistentValueContainer} class.
+ * API for storing persistent key-value data with debounced sync using a specified storage key.
+ * Storage synchronization is described in the {@link PersistentValueContainer} class.
  */
 export class ExtensionStorage<
     Data extends Record<string, unknown>,
     Key extends string = string,
 > {
     /**
-     * API for storing persistent value with debounced sync with the specified webextension storage key.
+     * API for storing persistent value with debounced sync using the specified storage key.
      */
     #container: PersistentValueContainer<Key, Data>;
 
     /**
      * Creates {@link ExtensionStorage} instance.
      * @param key The key to use for storing the data.
-     * @param api Webextension storage API.
+     * @param storage Storage interface implementation (replacing WebExtension API storage).
      */
     constructor(
         key: Key,
-        api: Storage.StorageArea,
+        storage: ExtendedStorageInterface<Key, Data, 'async'>, // Use StorageInterface for async storage handling
     ) {
-        this.#container = new PersistentValueContainer<Key, Data>(key, api);
+        this.#container = new PersistentValueContainer<Key, Data>(key, storage);
     }
 
     /**
@@ -39,9 +38,9 @@ export class ExtensionStorage<
 
     /**
      * Gets the value by the specified key.
-     * @param key The key to use for storing the value.
+     * @param key The key to retrieve the value.
      * @throws Error, if storage not initialized.
-     * @returns Data stored by the specified key.
+     * @returns The data stored by the specified key.
      */
     get<T extends keyof Data>(key: T): Data[T] {
         return this.#container.get()[key];
@@ -49,8 +48,8 @@ export class ExtensionStorage<
 
     /**
      * Sets the value by the specified key.
-     * @param key The key to use for storing the value.
-     * @param value New value.
+     * @param key The key for the value to be stored.
+     * @param value The new value.
      * @throws Error, if storage not initialized.
      */
     set<T extends keyof Data>(key: T, value: Data[T]): void {
@@ -61,7 +60,7 @@ export class ExtensionStorage<
 
     /**
      * Deletes the value by the specified key.
-     * @param key The key to use for storing the value.
+     * @param key The key for the value to be deleted.
      * @throws Error, if storage not initialized.
      */
     delete(key: keyof Data): void {